fix debugging add sweet allert and refresh data after succes deleted

+++ src/views/data/SumberAir.vue
@@ -2,22 +2,51 @@
 import { ref, onMounted } from 'vue'
 import { useLoading } from 'vue-loading-overlay'
 import MainWrapper from '@/components/MainWrapper.vue'
-import { getSumberAir, deleteSumberAir } from '@/lib/sumberAir';
+import { getSumberAir, deleteSumberAir } from '@/lib/sumberAir'
 import { getPertekData } from '@/lib/company.js'
-
+import Swal from 'sweetalert2'
 const $loading = useLoading()
 
 const sumberAir = ref([])
 const totalSumberAir = ref(0)
-
-const deleteData = async (id) => {
+const fetchData = async () => {
   const loader = $loading.show()
   try {
-    await deleteSumberAir(id)
+    sumberAir.value = await getSumberAir()
   } catch (e) {
-    console.error(e)
+    console.error('Error fetching data:', e)
+    Swal.fire('Error', 'Gagal mengambil data pencemaran air.', 'error')
   } finally {
     loader.hide()
+  }
+}
+const deleteData = async id => {
+  const { isConfirmed } = await Swal.fire({
+    title: 'Apa kamu yakin ?',
+    text: 'Kamu tidak akan bisa mengembalikan ini!',
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonText: 'Ya',
+    cancelButtonText: 'Tidak',
+    customClass: {
+      confirmButton: 'btn btn-primary',
+      cancelButton: 'btn btn-secondary',
+    },
+    buttonsStyling: false,
+  })
+
+  if (isConfirmed) {
+    const loader = $loading.show()
+
+    try {
+      await deleteSumberAir(id)
+      await fetchData()
+      Swal.fire('Deleted!', 'Data berhasil dihapus.', 'success')
+    } catch (e) {
+      console.error(e)
+    } finally {
+      loader.hide()
+    }
   }
 }
 
@@ -58,12 +87,17 @@ onMounted(async () => {
         <div class="content-page-header content-page-headersplit mb-2">
           <div>
             <h3>Data Sumber Air</h3>
-            <p>Anda memasukan jumlah sumber air sebanyak {{ totalSumberAir }}</p>
+            <p>
+              Anda memasukan jumlah sumber air sebanyak {{ totalSumberAir }}
+            </p>
           </div>
           <div class="list-btn">
             <ul>
               <li>
-                <router-link class="btn btn-primary" to="/Data/SumberAir/Tambah">
+                <router-link
+                  class="btn btn-primary"
+                  to="/Data/SumberAir/Tambah"
+                >
                   <!-- Change route as necessary -->
                   <i class="fa fa-plus me-2"></i>Tambah Sumber Air
                 </router-link>
@@ -93,10 +127,12 @@ onMounted(async () => {
                     <td>
                       <router-link
                         :to="{
-                          path: data.id ? `/Data/SumberAir/Edit/${data.id}` : '/Data/SumberAir/Tambah',
+                          path: data.id
+                            ? `/Data/SumberAir/Edit/${data.id}`
+                            : '/Data/SumberAir/Tambah',
                           query: {
-                            jenis: data.id ? undefined : data.jenis
-                          }
+                            jenis: data.id ? undefined : data.jenis,
+                          },
                         }"
                         class="btn btn-primary"
                         >Edit</router-link
